+++ src/templates/PageLayout/index.ts
@@ -1,0 +1,2 @@
+import PageLayout from './PageLayout'
+export default PageLayout
